Minor Fixes

+++ main.py
@@ -31,11 +31,18 @@ class User(UserMixin):
     @staticmethod
     def get(user_id):
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("SELECT * FROM Persons WHERE id = %s",(user_id,))
+        cursor.execute("SELECT * FROM persons WHERE id = %s",(user_id,))
         acc = cursor.fetchone()
         if acc:
             return User(id=acc['id'],username=acc['username'],role=acc['role'])
         return None
+
+def is_admin():
+    try:
+        admin = current_user.is_authenticated and current_user.role == "admin"
+    except:
+        admin = False
+    return admin
 
 @app.route('/logout')
 @login_required
@@ -46,17 +53,16 @@ def logout():
 @app.route('/tasks',methods=['GET'])
 @login_required
 def view_task():
-    admin = current_user.role == "admin"
     user_id = current_user.id
     user = current_user.username
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     cursor.execute("SELECT * FROM tasks WHERE user_id = %s",(user_id,))
-    if admin:
+    if is_admin():
         cursor.execute("SELECT * FROM tasks")
     tasks = cursor.fetchall()
     users = []
     l = []
-    if admin:
+    if is_admin():
         for task in tasks:
             cursor.execute("SELECT * FROM persons WHERE id = %s",(task['user_id'],))
             user = cursor.fetchone()
@@ -77,50 +83,47 @@ def view_task():
 @login_required
 def add_task():
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-    admin = current_user.role == "admin"
     if "task" in request.form and "time" in request.form and request.method == 'POST':
         task = request.form['task']
         time = request.form['time']
         user_id = current_user.id
-        if admin:
-            user_id = request.form['userid']
-            print(user_id)
+        if is_admin():
+            user_id = int(request.form['userid'])
         
-        cursor.execute("INSERT INTO Tasks (user_id, time, task) VALUES (%s, %s, %s)",(user_id,time,task))
+        cursor.execute("INSERT INTO tasks (user_id, time, task) VALUES (%s, %s, %s)",(user_id,time,task))
         mysql.connection.commit()
         return redirect(url_for('view_task'))  
-    admin = current_user.role == "admin"
-    cursor.execute("SELECT * FROM Persons")
+    cursor.execute("SELECT * FROM persons")
     users = cursor.fetchall()
-    return render_template("add_tasks.html",admin=admin,users=users)
+    return render_template("add_tasks.html",admin=is_admin(),users=users)
 
 @app.route('/delete_task/<int:task_id>',methods=['GET','POST'])
 @login_required
 def delete_task(task_id):
-    admin = current_user.role == "admin"
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-    cursor.execute("DELETE FROM TASKS WHERE id = %s and user_id = %s",(task_id,current_user.id))
-    if admin:
+
+    if is_admin():
         cursor.execute("DELETE FROM TASKS WHERE id = %s",(task_id,))
+    else:
+        cursor.execute("DELETE FROM TASKS WHERE id = %s and user_id = %s",(task_id,current_user.id))
+
     mysql.connection.commit()
     return redirect(url_for('view_task'))
 
 @app.route('/update_task/<int:task_id>',methods=['GET','POST'])
 @login_required
 def update_task(task_id):
-    admin = current_user.role == "admin"
     cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     if request.method == 'POST':
-        cursor.execute("UPDATE TASKS SET time = %s, task = %s WHERE id = %s AND user_id = %s",(request.form['time'], request.form['task'], task_id, current_user.id))
-
-        if admin:
+        if is_admin():
             cursor.execute("UPDATE TASKS SET time = %s, task = %s WHERE id = %s",(request.form['time'], request.form['task'], task_id))
-        
+        else:
+            cursor.execute("UPDATE TASKS SET time = %s, task = %s WHERE id = %s AND user_id = %s",(request.form['time'], request.form['task'], task_id, current_user.id))
         mysql.connection.commit()
         return redirect(url_for('view_task'))
     
     cursor.execute("SELECT * FROM TASKS WHERE user_id = %s and id = %s",(current_user.id,task_id))
-    if admin:
+    if is_admin():
         cursor.execute("SELECT * FROM TASKS WHERE id = %s",(task_id,))
     task = cursor.fetchone()
     return render_template('update_task.html',task_id=task_id,task=task)
@@ -128,11 +131,7 @@ def update_task(task_id):
 @app.route('/',methods=['GET'])
 def home():
     is_user = current_user.is_authenticated
-    try:
-        admin = current_user.role == "admin"
-    except:
-        admin = False
-    return render_template("index.html",is_user=is_user,admin=admin,user=current_user)
+    return render_template("index.html",is_user=is_user,admin=is_admin(),user=current_user)
 
 @login_manager.user_loader
 def load_user(user_id):
@@ -146,7 +145,7 @@ def login():
         password = request.form['password']
         msg=""
 
-        cursor.execute("SELECT * FROM Persons WHERE USERNAME = %s",(username,))
+        cursor.execute("SELECT * FROM persons WHERE USERNAME = %s",(username,))
         acc = cursor.fetchone()
         if acc:
             is_valid = bcrypt.check_password_hash(acc['password'], password)
@@ -165,10 +164,14 @@ def login():
                 rd = "login"
                 img = "error.png"
                 color = "#fc4747"
-            return render_template("result.html",title=title,msg=msg,rd=rd,image=img,color=color)
         else:
-            msg = "Your Account doesnt exist! Kindly register first!"
-        mysql.connection.commit()
+            title = "Your Account doesnt exist!"
+            msg = "The Login Failed as your account doesn't exist in our database. Kindly register first."
+            rd = "login"
+            img = "error.png"
+            color = "#fc4747"
+        
+        return render_template("result.html",title=title,msg=msg,rd=rd,image=img,color=color)
     return render_template("login.html")
 
 @app.route('/register',methods=['GET','POST'])
@@ -179,7 +182,7 @@ def register():
         password = request.form['password']
         city = request.form['city']
         role = "user"
-        cursor.execute("SELECT * FROM Persons WHERE USERNAME = %s",(username, ))
+        cursor.execute("SELECT * FROM persons WHERE USERNAME = %s",(username, ))
         acc = cursor.fetchone()
 
         hashed_password = bcrypt.generate_password_hash(password).decode('utf-8')
@@ -187,7 +190,7 @@ def register():
         if acc:
             title = "Account already Exists!" 
         else:
-            cursor.execute("INSERT INTO Persons (username, password, city,role) VALUES (%s, %s, %s,%s)",(username,hashed_password,city,role))
+            cursor.execute("INSERT INTO persons (username, password, city,role) VALUES (%s, %s, %s,%s)",(username,hashed_password,city,role))
             title = "You have successfully registered!"
         mysql.connection.commit()
         
@@ -198,26 +201,21 @@ def register():
 @login_required
 @app.route('/users',methods=['GET'])
 def view_users():
-    try:
-        admin = current_user.role == "admin"
-    except:
-        admin = False
-    if admin:
+    if is_admin():
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("SELECT * FROM Persons")
+        cursor.execute("SELECT * FROM persons")
         users = cursor.fetchall()
-        return render_template("users.html",admin=admin,users=users)
+        return render_template("users.html",admin=is_admin(),users=users)
     else:
         return render_template("result.html",title="Unauthorized Access!",msg="The following site can only be accessed by an admin. Contact your administrator for more information.",color="#fc4747",image="error.png",rd="home")
 
 @login_required
 @app.route('/change_role/<int:user_id>',methods=['GET','POST'])
 def change_role(user_id):
-    admin = current_user.role == "admin"
-    if admin:
+    if is_admin():
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
     
-        cursor.execute("SELECT * FROM Persons WHERE id = %s",(user_id,))
+        cursor.execute("SELECT * FROM persons WHERE id = %s",(user_id,))
         user = cursor.fetchone()
 
         if user['role'] == "admin":
@@ -228,16 +226,16 @@ def change_role(user_id):
         cursor.execute("UPDATE persons SET role = %s WHERE id = %s",(role,user_id,))
         mysql.connection.commit()
         return redirect(url_for('view_users'))
+    
     return render_template("result.html",title="Unauthorized Access!",msg="The following site can only be accessed by an admin. Contact your administrator for more information.",color="#fc4747",image="error.png",rd="home")
 
 @login_required
 @app.route('/delete_user/<int:user_id>',methods=['GET','POST'])
 def delete_user(user_id):
-    admin = current_user.role == "admin"
-    if admin:
+    if is_admin():
         cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        cursor.execute("DELETE FROM Persons WHERE id =  %s",(user_id,))
-        cursor.execute("DELETE FROM Tasks WHERE user_id = %s",(user_id,))
+        cursor.execute("DELETE FROM persons WHERE id =  %s",(user_id,))
+        cursor.execute("DELETE FROM tasks WHERE user_id = %s",(user_id,))
 
         mysql.connection.commit()
         return redirect(url_for('view_users'))
